Table hunt: Tweak score handling.
Previously we'd discard any table that had an 'awkwardness
score' of more than 0.3.

Relax that requirement when we are given an bbox to search in
from outside. The expectation, in this case, is that some
other mechanism has identified this as being likely to be
a table. So we don't need to be so conservative.