Texas chainsearch massacre!
The search code must match the spun and unspun text positions of the
search result with the corresponding fz_stext_char positions.

Rejig search code to use a flattened array of stext chars instead of
(with extreme difficulty) iteratively stepping through the stext
structure in lockstep with the same behavior as do_flatten.

Also improve the mutool grep tool and rewrite it to use a different
approach to walking the results (using the new fz_stext_char parent
pointers).