Fix formatting of image dimensions, causing warning in wasm.
The image dimensions originates from a fz_pixmap where they are of type int.
They dimensions are stored in sh/sw which are of type affint, and alias for
int64_t. The code then used to cast sh to int64_t in order to format them with
fz_format_string("%ld") where "%ld" formats a 64-bit signed integer.

The problem is that fz_warn() has printf-like warnings, so it "knows" that
"%ld" formats a long, not an int64_t causing a warning. Since the dimensions
are originally int, here we just cast to that and format with "%d" avoiding
the warning.