Rework sanitisation of Text operators.
There are multiple different operators involved in the alteration
of the Text Matrix.

Tm sets it directly, Td adjusts it slightly, T* adjusts it relative
by an amount previously set by TL, etc.

Previously, we'd attempted to pick which one to use based on
some hairy Tm_pending and Td_pending flags, but this had failed
to help with leading etc.

Here we rework the system just to keep the desired Tm. At the point
at which we want to flush the changes we now examine the Tm
compared to the previously sent one, and pick some combination of
operators to send it as efficiently as possible.

Getting this absolutely correct is tricky as, theoretically, we
might be better off in some cases leaving Leading set to an old
value, rather than resetting it each time. But that's going to
be impossible to solve in a 'single-pass' fashion. What we have
should be pretty good.