Bug 708791: XPS: cope with 'untidy' split _rels/.rels pieces.
Gather the pieces into a single buffer before parsing.

We had some existing code in the XPS interpreter that did this
gathering for us, but it relied on the document having been
opened. Here, we refactor the code to give us a common routine
we can use.

Improve that code to cope with both '/' and '\\' separators.

And improve the recogniser to spot OXPS files too.